refactor(redis_ops): extract orchestration module
Extract high-level dump orchestration into orchestration.rs:
- dump_keys: Public function to dump a set of keys with parallel workers
- run_dump: Public entry point for the complete dump process
- log_processing_start, log_worker_distribution, report_completion: Logging helpers

mod.rs is now a clean module hub with only declarations and re-exports.
